+++ lib/packages.dart
@@ -5,6 +5,7 @@ export 'package:kabutor/firebase_options.dart';
 export 'package:flutter_screenutil/flutter_screenutil.dart';
 export 'package:get/get.dart';
 export 'package:google_fonts/google_fonts.dart';
+export 'package:flutter_otp_text_field/flutter_otp_text_field.dart';
 
 export 'utils/theme/palette.dart';
 export 'ui/shared/widgets/kabutor_logo.dart';
@@ -13,7 +14,7 @@ export '/ui/shared/widgets/custom_button.dart';
 export 'ui/shared/widgets/Sscaffold_backgound.dart';
 export 'ui/pages/auth/signin.dart';
 export 'package:kabutor/ui/pages/auth/forgotpass.dart';
-export 'ui/pages/auth/verifacition_page.dart';
+export 'ui/pages/auth/verifacition_forgot_pass_page.dart';
 export 'ui/pages/auth/newaccount_page.dart';
 export 'ui/pages/auth/signin_pass_authscreen.dart';
