+++ lib/ui/shared/widgets/Sscaffold_backgound.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: file_names
+
 import '../../../packages.dart';
 
 class ScaffoldBackgound extends StatelessWidget {
