+++ pubspec.yaml
@@ -23,6 +23,7 @@ dependencies:
   get: ^4.6.5
   flutter_screenutil: ^5.6.0
   google_fonts: ^3.0.1
+  flutter_otp_text_field: ^1.1.1
   
 
 dev_dependencies:
